langchain_project/requirements.txt
langchain
langchain-google-genai
python-dotenv
uv
langchain-community
wikipedia
langchain-experimental
pandas
beautifulsoup4
lxml
yfinance